+++ functions-python/process_validation_report/src/main.py
@@ -327,3 +327,55 @@ def process_validation_report(request):
         f"Processing validation report version {validator_version} for dataset {dataset_id} in feed {feed_id}."
     )
     return create_validation_report_entities(feed_id, dataset_id, validator_version)
+
+
+@functions_framework.http
+def compute_validation_report_counters(session: sqlalchemy.orm.Session):
+    """
+    Compute the total number of errors, warnings, and info notices,
+    as well as the number of distinct codes for each severity level
+    across all validation reports in the database, and write the results to the database.
+
+    :param session: The database session
+    """
+    db = Database()
+    with db.start_db_session(echo=False) as session:
+        # Query all validation reports
+        validation_reports = session.query(Validationreport).all()
+
+        for report in validation_reports:
+            # Initialize counters for the current report
+            total_info, total_warning, total_error = 0, 0, 0
+            info_codes, warning_codes, error_codes = set(), set(), set()
+
+            # Process associated notices
+            for notice in report.notices:
+                match notice.severity:
+                    case "INFO":
+                        total_info += notice.total_notices
+                        info_codes.add(notice.notice_code)
+                    case "WARNING":
+                        total_warning += notice.total_notices
+                        warning_codes.add(notice.notice_code)
+                    case "ERROR":
+                        total_error += notice.total_notices
+                        error_codes.add(notice.notice_code)
+                    case _:
+                        logging.warning(f"Unknown severity: {notice.severity}")
+
+            # Update the report with computed counters
+            report.total_info = total_info
+            report.total_warning = total_warning
+            report.total_error = total_error
+            report.unique_info_count = len(info_codes)
+            report.unique_warning_count = len(warning_codes)
+            report.unique_error_count = len(error_codes)
+
+            logging.info(
+                f"Updated ValidationReport {report.id} with counters: "
+                f"INFO={total_info}, WARNING={total_warning}, ERROR={total_error}, "
+                f"Unique INFO Codes={len(info_codes)}, Unique WARNING Codes={len(warning_codes)}, "
+                f"Unique ERROR Codes={len(error_codes)}"
+            )
+
+    return {"message": "Validation report counters computed successfully."}, 200
